refactor: [회원가입] 파일 경로 수정
