+++ src/app/(whitebg)/users/join/_components/JoinAgreement.tsx
