+++ src/app/(whitebg)/users/join/_components/JoinCompleted.module.css
