+++ src/app/(whitebg)/users/join/_components/JoinReceiveNotification.tsx
