+++ src/app/(whitebg)/users/join/page.tsx
@@ -1,14 +1,14 @@
 'use client';
 import React, {useEffect} from 'react';
-import JoinAgreement from '@/app/(whitebg)/users/_components/join/JoinAgreement';
-import JoinReceiveNotification from '@/app/(whitebg)/users/_components/join/JoinReceiveNotification';
-import JoinEmailVerification from '@/app/(whitebg)/users/_components/join/JoinEmailVerification';
-import JoinExperience from '@/app/(whitebg)/users/_components/join/JoinExperience';
-import JoinCompleted from '@/app/(whitebg)/users/_components/join/JoinCompleted';
+import JoinAgreement from '@/app/(whitebg)/users/join/_components/JoinAgreement';
+import JoinReceiveNotification from '@/app/(whitebg)/users/join/_components/JoinReceiveNotification';
+import JoinEmailVerification from '@/app/(whitebg)/users/join/_components/JoinEmailVerification';
+import JoinExperience from '@/app/(whitebg)/users/join/_components/JoinExperience';
+import JoinCompleted from '@/app/(whitebg)/users/join/_components/JoinCompleted';
 import {useUserJoin} from '@/hooks/queries/useUserJoin';
 import {useRouter} from 'next/navigation';
 import {useUserJoinContext} from '@/contexts/UserJoinProvider';
-import JoinPageTitle from '@/app/(whitebg)/users/_components/join/JoinPageTitle';
+import JoinPageTitle from '@/app/(whitebg)/users/join/_components/JoinPageTitle';
 import styles from './page.module.css';
 import Button from '@/components/Button';
 import ArrowLeft from '@/assets/icon/arrow-left.svg';
